Fix buckets in social weather not shown correctly
Previously the one-day timeframe will not be showing correct buckets
owing to timezone. For instance, viewing the page on 4PM would result
in returning the data from 8AM (of the previous day) to 8AM (of the
current day). This is owing to the timezone is not selected correctly
and hence the wrong period of time is being picked.